cabal2arch.cabal: advertise the location of the Git repository
Unfortunately, the 'source-repository' tag is supported only in Cabal 1.6 or
later, and older versions of Cabal will not be able to parse this file anymore
because of the exotic syntax. It feels alright, though, to expect people to
have a recent version of Cabal, especially since this package is intended to be
used by people running ArchLinux.